% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit.
}
\examples{
C_to_F(0)
C_to_F(c(-10, 0, 25))
}
