% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in Kelvin
}
\description{
Convert degrees Celsius temperatures to Kelvin.
}
\examples{
C_to_K(0)
C_to_K(c(-10, 0, 100))
}
