% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius.
}
\examples{
F_to_C(32)
F_to_C(c(32, 50, 77))
}
