% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in Kelvin
}
\description{
Convert degrees Fahrenheit temperatures to Kelvin.
}
\examples{
F_to_K(32)
F_to_K(c(32, 212))
}
