% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to Celsius}
\usage{
K_to_C(K_temp)
}
\arguments{
\item{K_temp}{The temperature in Kelvin}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert temperatures from Kelvin to degrees Celsius.
}
\examples{
K_to_C(273.15)
K_to_C(c(273.15, 310.15))
}
