% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvJB_functions.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{The temperature in Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert temperatures from Kelvin to degrees Fahrenheit.
}
\examples{
K_to_F(273.15)
K_to_F(c(273.15, 310.15))
}
